refactor(config): migração das camadas core para Spring Boot + Spring Data JPA

+++ src/main/java/br/com/picarauto/model/IItemServicoOS.java
@@ -11,6 +11,6 @@ package br.com.picarauto.model;
  * @author Caio4breu
  */
 public interface IItemServicoOS {
-    Integer getId();
+    Long getId();
     String getDescricao();
 }
